--- .vscode/tasks.json
@@ -1,25 +1,0 @@
-{
-  "tasks": [
-    {
-      "type": "cppbuild",
-      "label": "C/C++: g++.exe build active file",
-      "command": "C:\\mingw64\\bin\\g++.exe",
-      "args": [
-        "-g",
-        "${workspaceFolder}\\*.cpp",
-        "-o",
-        "${fileDirname}\\${fileBasenameNoExtension}.exe"
-      ],
-      "options": {
-        "cwd": "${fileDirname}"
-      },
-      "problemMatcher": ["$gcc"],
-      "group": {
-        "kind": "build",
-        "isDefault": true
-      },
-      "detail": "Task generated by Debugger."
-    }
-  ],
-  "version": "2.0.0"
-}
