+++ House.cpp
@@ -2,12 +2,14 @@
 #include <vector>
 #include <sstream>
 #include "House.h"
+#include <stdlib.h>
+#include <string.h>
 using std::cin;
 using std::cout;
 using std::string;
 using std::vector;
 using std::stringstream;
-House::House(){};
+using std::end;
 House::House(string location, string description){
     this->location = location;
     this->description = description;
@@ -93,7 +95,7 @@ void House::set_info(double houserate,double min_occ_rate,
 void House::show_ratings(){
     int i =1;
     cout << "House ratings: " << std::endl;
-    for(Rating r: this->ratings){
+    for (Rating r: this->ratings) {
         cout << i << ". Score: " << r.getScore() << " || Comment: " << r.getComment() << std::endl;
         i++; 
     }
