+++ Rating.h
@@ -2,10 +2,13 @@
 #define RATING_H
 #include <iostream>
 using std::string;
+// g++ Date.cpp Global.cpp House.cpp Login.cpp MainPage.cpp Member.cpp
 
 class Rating{
+    
     double score;
     string comment;
+
     public:
     Rating();
     Rating(int score, string comment);
